Split sql core from default tables
